Moves the cli commands to a module
Moves all of the cli commands from a single file to a file-per-command
in the ckan.lib.commands module.

Unfortunately removing cli.py from the pep whitelist has made the pep8
command now start checking a lot of files that it didn't before - I
suspect a bug in the pep8 test somewhere.  As such, this is a WIP.
