+++ ckan/lib/commands/notification.py
@@ -1,0 +1,35 @@
+from ckan.lib.commands import CkanCommand
+
+
+class Notification(CkanCommand):
+    '''Send out modification notifications.
+
+    In "replay" mode, an update signal is sent for each dataset in
+    the database.
+
+    Usage:
+      notify replay           - send out modification signals
+    '''
+
+    summary = __doc__.split('\n')[0]
+    usage = __doc__
+    max_args = 1
+    min_args = 0
+
+    def command(self):
+        self._load_config()
+        from ckan.model import Session, Package, DomainObjectOperation
+        from ckan.model.modification import DomainObjectModificationExtension
+
+        if not self.args:
+            # default to run
+            cmd = 'replay'
+        else:
+            cmd = self.args[0]
+
+        if cmd == 'replay':
+            dome = DomainObjectModificationExtension()
+            for package in Session.query(Package):
+                dome.notify(package, DomainObjectOperation.changed)
+        else:
+            print 'Command %s not recognized' % cmd
