+++ ckan/lib/commands/ratings.py
@@ -1,0 +1,48 @@
+from ckan.lib.commands import CkanCommand
+
+
+class Ratings(CkanCommand):
+    '''Manage the ratings stored in the db
+
+    Usage:
+      ratings count                 - counts ratings
+      ratings clean                 - remove all ratings
+      ratings clean-anonymous       - remove only anonymous ratings
+    '''
+
+    summary = __doc__.split('\n')[0]
+    usage = __doc__
+    max_args = 1
+    min_args = 1
+
+    def command(self):
+        self._load_config()
+
+        cmd = self.args[0]
+        if cmd == 'count':
+            self.count()
+        elif cmd == 'clean':
+            self.clean()
+        elif cmd == 'clean-anonymous':
+            self.clean(user_ratings=False)
+        else:
+            print 'Command %s not recognized' % cmd
+
+    def count(self):
+        import ckan.model as model
+        q = model.Session.query(model.Rating)
+        print "%i ratings" % q.count()
+        q = q.filter(model.Rating.user_id == None)
+        print "of which %i are anonymous ratings" % q.count()
+
+    def clean(self, user_ratings=True):
+        import ckan.model as model
+        q = model.Session.query(model.Rating)
+        print "%i ratings" % q.count()
+        if not user_ratings:
+            q = q.filter(model.Rating.user_id == None)
+            print "of which %i are anonymous ratings" % q.count()
+        ratings = q.all()
+        for rating in ratings:
+            rating.purge()
+        model.repo.commit_and_remove()
