+++ ckan/tests/test_coding_standards.py
@@ -409,7 +409,9 @@ class TestPep8(object):
         'ckan/lib/alphabet_paginate.py',
         'ckan/lib/app_globals.py',
         'ckan/lib/captcha.py',
-        'ckan/lib/cli.py',
+        'ckan/lib/commands/search.py',
+        'ckan/lib/commands/views.py',
+        'ckan/lib/commands/ratings.py',
         'ckan/lib/create_test_data.py',
         'ckan/lib/dictization/__init__.py',
         'ckan/lib/dictization/model_dictize.py',
