+++ setup.py
@@ -19,29 +19,29 @@ entry_points = {
         'main = ckan.config.install:CKANInstaller',
     ],
     'paste.paster_command': [
-        'db = ckan.lib.cli:ManageDb',
-        'create-test-data = ckan.lib.cli:CreateTestDataCommand',
-        'sysadmin = ckan.lib.cli:Sysadmin',
-        'user = ckan.lib.cli:UserCmd',
-        'dataset = ckan.lib.cli:DatasetCmd',
-        'search-index = ckan.lib.cli:SearchIndexCommand',
-        'ratings = ckan.lib.cli:Ratings',
-        'notify = ckan.lib.cli:Notification',
-        'celeryd = ckan.lib.cli:Celery',
-        'rdf-export = ckan.lib.cli:RDFExport',
-        'tracking = ckan.lib.cli:Tracking',
-        'plugin-info = ckan.lib.cli:PluginInfo',
-        'profile = ckan.lib.cli:Profile',
-        'color = ckan.lib.cli:CreateColorSchemeCommand',
+        'db = ckan.lib.commands.db:ManageDb',
+        'create-test-data = ckan.lib.commands.test_data:CreateTestDataCommand',
+        'sysadmin = ckan.lib.commands.sysadmin:Sysadmin',
+        'user = ckan.lib.commands.user:UserCmd',
+        'dataset = ckan.lib.commands.dataset:DatasetCmd',
+        'search-index = ckan.lib.commands.search:SearchIndexCommand',
+        'ratings = ckan.lib.commands.ratings:Ratings',
+        'notify = ckan.lib.commands.notification:Notification',
+        'celeryd = ckan.lib.commands.celery:Celery',
+        'rdf-export = ckan.lib.commands.rdf:RDFExport',
+        'tracking = ckan.lib.commands.tracking:Tracking',
+        'plugin-info = ckan.lib.commands.plugin_info:PluginInfo',
+        'profile = ckan.lib.commands.profile:Profile',
+        'color = ckan.lib.commands.color_scheme:CreateColorSchemeCommand',
         'check-po-files = ckan.i18n.check_po_files:CheckPoFiles',
-        'trans = ckan.lib.cli:TranslationsCommand',
-        'minify = ckan.lib.cli:MinifyCommand',
-        'less = ckan.lib.cli:LessCommand',
+        'trans = ckan.lib.commands.translations:TranslationsCommand',
+        'minify = ckan.lib.commands.minify:MinifyCommand',
+        'less = ckan.lib.commands.less:LessCommand',
         'datastore = ckanext.datastore.commands:SetupDatastoreCommand',
         'datapusher = ckanext.datapusher.cli:DatapusherCommand',
-        'front-end-build = ckan.lib.cli:FrontEndBuildCommand',
-        'views = ckan.lib.cli:ViewsCommand',
-        'config-tool = ckan.lib.cli:ConfigToolCommand',
+        'front-end-build = ckan.lib.commands.front_end:FrontEndBuildCommand',
+        'views = ckan.lib.commands.views:ViewsCommand',
+        'config-tool = ckan.lib.commands.config:ConfigToolCommand',
     ],
     'console_scripts': [
         'ckan-admin = bin.ckan_admin:Command',
